% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gfdiff.R
\name{Gfdiff}
\alias{Gfdiff}
\title{Generalized differencing methods}
\usage{
Gfdiff(y, x)
}
\arguments{
\item{y}{series name}

\item{x}{series name,}
}
\description{
This function uses generalized differencing method for correction autocorrelation.
}
\examples{
IHR = REcoData$IHR
ITH = REcoData$ITH
Gfdiff(IHR, ITH)

}
\keyword{Autocorrelation}
