% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REcoData.R
\docType{data}
\name{REcoData}
\alias{REcoData}
\title{REcoData}
\format{A data frame containing :}
\usage{
REcoData
}
\description{
Mothly time series data between 2010.1-2019.4
}
\examples{
summary(REcoData)
}
\keyword{datasets}
