% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REcoData_Panel.R
\docType{data}
\name{REcoData_Panel}
\alias{REcoData_Panel}
\title{REcoData_Panel}
\format{A data frame containing :}
\usage{
REcoData_Panel
}
\description{
Panel data between 1996-2017 for G8 countries
}
\examples{
summary(REcoData_Panel)
}
\keyword{datasets}
