% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REcoData_Panel_UR.R
\docType{data}
\name{REcoData_Panel_UR}
\alias{REcoData_Panel_UR}
\title{REcoData_Panel_UR}
\format{A data frame containing :}
\usage{
REcoData_Panel_UR
}
\description{
Panel data between 1980-2017 for fifteen countries
}
\examples{
summary(REcoData_Panel_UR)
}
\keyword{datasets}
