% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REcoData_Tourism.R
\docType{data}
\name{REcoData_Tourism}
\alias{REcoData_Tourism}
\title{REcoData_Tourism}
\format{A data frame containing :}
\usage{
REcoData_Tourism
}
\description{
Quarterly tourism revenue data for Turkey between 2003.Q1-2019.Q2
}
\examples{
summary(REcoData_Tourism)
}
\keyword{datasets}
