% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResTest.R
\name{ResTest}
\alias{ResTest}
\title{Restriction Tests}
\usage{
ResTest(y, x1, x2)
}
\arguments{
\item{y}{series name,}

\item{x1}{series name}

\item{x2}{series name}
}
\description{
This function computes LM, LR and Wald test statistics for redundant variable.
}
\examples{
IHR = REcoData$IHR
ITH = REcoData$ITH
DK =REcoData$DK
ResTest(IHR,ITH,DK)

}
\keyword{Restriction}
\keyword{Tests}
