% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HausmanTest.R
\name{HausmanTest}
\alias{HausmanTest}
\title{Hausmann Test for identification}
\usage{
HausmanTest(y, x, z)
}
\arguments{
\item{y}{series name}

\item{x}{series name,}

\item{z}{series name}
}
\description{
This function allows you to make Hausman Test for identification
}
\examples{
IHR = REcoData$IHR
ITH = REcoData$ITH
DK =REcoData$DK
HausmanTest(IHR,ITH,DK)

}
\references{
Selahattin Güriş, Ebru Çağlayan Akay, Burak Güriş, R ile Temel Ekonometri, DER Yayinevi, 2020.
}
\keyword{restriction}
