% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REcoData_DCM.R
\docType{data}
\name{REcoData_DCM}
\alias{REcoData_DCM}
\title{REcoData_DCM}
\format{
A data frame containing :

Yoksulluk Poverty, 1 if poor, 0 if not poor

YKGelir Annual Disposable Income

KK 1 if living in the city, 0 if living in the countryside
}
\usage{
REcoData_DCM
}
\description{
Poverty data for 100 people
}
\examples{
summary(REcoData_DCM)
}
\keyword{datasets}
