% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REcoData_SEM.R
\docType{data}
\name{REcoData_SEM}
\alias{REcoData_SEM}
\title{REcoData_SEM}
\format{
A data frame containing :

LIH Natural Logarithm of Import

LIT Natural Logarithm of Export

LPA Natural Logarithm of Money Supply

LDK Natural Logarithm of Exchange Rate
}
\usage{
REcoData_SEM
}
\description{
Yearly data for Turkey between 1990-2002
}
\examples{
summary(REcoData_SEM)
}
\keyword{datasets}
