% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdreg.R
\name{stdreg}
\alias{stdreg}
\title{Standardized Regression}
\usage{
stdreg(y, x)
}
\arguments{
\item{y}{series name,}

\item{x}{series name}
}
\description{
This function computes standardized regression model.
}
\examples{
IHR = REcoData$IHR
ITH = REcoData$ITH
stdreg(IHR,ITH)

}
\references{
Selahattin Güriş, Ebru Çağlayan Akay, Burak Güriş, R ile Temel Ekonometri, DER Yayinevi, 2020.
}
\keyword{Regression}
\keyword{Standardized}
