% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wls.R
\name{Wls}
\alias{Wls}
\title{Weighted Least Square}
\usage{
Wls(y, x)
}
\arguments{
\item{y}{series name,}

\item{x}{series name}
}
\description{
This Function makes Weighted Least Square estimation.
}
\examples{
IHR = REcoData$IHR
ITH = REcoData$ITH
Wls(ITH,IHR)

}
\references{
Selahattin Güriş, Ebru Çağlayan Akay, Burak Güriş, R ile Temel Ekonometri, DER Yayinevi, 2020.
}
\keyword{Least}
\keyword{Square}
\keyword{Weighted}
\keyword{estimation}
